go.mod: don't set non-zero patch version on go directive
The go directive indicates which language features are used. A specific
patch version is unnecessary because Golang language features don't
change between patch versions.

Downstream packages that import go.opentelemetry.io/ebpf-profiler need
to use a version equal or greater than the one specified in the go
directive. Keep the patch version to zero to avoid unnecessary
constraints.

It is possible to use the toolchain directive if we want to enforce that
ebpf-profiler is built with a more recent Go version than specified in
the go directive (e.g. for critical security patches) but as far as I
know, this is not the case here. The toolchain directive would be used
for building go.opentelemetry.io/ebpf-profiler as main module but would
be ignored by downstream packages.

Signed-off-by: Alban Crequy <alban.crequy@gmail.com>